Renamed deprecatedContractsAPI back to its original name
