+++ examples/nodejs/src/escrow-flow.ts
@@ -88,17 +88,14 @@ async function main(action: "buy" | "sell", otherAddress: string, amount: number
   console.log("Mediator: " + Mediator);
   console.log("Amount: " + amount);
 
-  const [contractId, txId] = await runtime.deprecatedContractAPI.createContract(
-    {
-      contract: escrow,
-      roles: { Buyer, Seller, Mediator },
-    }
-  );
+  const contractInstance = await runtime.newContractAPI.createContract({
+    contract: escrow,
+    roles: { Buyer, Seller, Mediator },
+  });
 
-  console.log("Contract ID: " + contractId);
-  console.log("Transaction ID: " + txId);
+  console.log("Contract ID: " + contractInstance.contractId);
 
   console.log("Waiting for confirmation...");
-  await wallet.waitConfirmation(txId);
+  await contractInstance.waitForConfirmation();
   console.log("Contract created successfully!");
 }
