+++ packages/runtime/lifecycle/src/generic/contracts.ts
