+++ packages/runtime/lifecycle/src/generic/runtime.ts
@@ -4,7 +4,7 @@ import { WalletAPI } from "@marlowe.io/wallet/api";
 import { FPTSRestAPI, RestClient } from "@marlowe.io/runtime-rest-client";
 
 import { mkPayoutLifecycle } from "./payouts.js";
-import { mkContractLifecycle } from "./deprecated-contracts.js";
+import { mkContractLifecycle } from "./contracts.js";
 import { mkApplicableActionsAPI } from "./applicable-actions.js";
 import * as NewContract from "./new-contract-api.js";
 
@@ -13,15 +13,11 @@ export function mkRuntimeLifecycle(
   restClient: RestClient,
   wallet: WalletAPI
 ): RuntimeLifecycle {
-  const deprecatedContractAPI = mkContractLifecycle(
-    wallet,
-    deprecatedRestAPI,
-    restClient
-  );
+  const contracts = mkContractLifecycle(wallet, deprecatedRestAPI, restClient);
   return {
     wallet: wallet,
     restClient,
-    deprecatedContractAPI,
+    contracts,
     newContractAPI: NewContract.mkContractsAPI({ wallet, restClient }),
     payouts: mkPayoutLifecycle(wallet, deprecatedRestAPI, restClient),
     applicableActions: mkApplicableActionsAPI({ restClient, wallet }),
